move code from lib.ts and types.ts into index.ts
